+++ lib/llm/active_record/acts_as_agent.rb
@@ -150,7 +150,7 @@ module LLM::ActiveRecord
       # @return [Hash]
       def resolve_options(option)
         case option
-        when Proc, Hash then resolve_option(option)
+        when Proc, Symbol, Hash then resolve_option(option)
         else ActsAsAgent::EMPTY_HASH.dup
         end
       end