+++ lib/llm/active_record/acts_as_llm.rb
@@ -270,7 +270,7 @@ module LLM::ActiveRecord
       # @return [Hash]
       def resolve_options(option)
         case option
-        when Proc, Hash then resolve_option(option)
+        when Proc, Symbol, Hash then resolve_option(option)
         else ActsAsLLM::EMPTY_HASH.dup
         end
       end